+++ packages/widgets/applications-portal-widget/CHANGELOG.md
@@ -2,6 +2,16 @@
 
 This file was generated using [@jscutlery/semver](https://github.com/jscutlery/semver).
 
+## [0.1.0](https://github.com/descope/descope-js/compare/applications-portal-widget-0.0.2...applications-portal-widget-0.1.0) (2024-10-14)
+
+### Dependency Updates
+
+* `web-js-sdk` updated to version `1.18.0`
+
+### Features
+
+* apps portal sdks ([#808](https://github.com/descope/descope-js/issues/808)) ([30b11b0](https://github.com/descope/descope-js/commit/30b11b0ec8252281ed3cfb273e415edfa2fa1070))
+
 ## [0.0.2](https://github.com/descope/descope-js/compare/applications-portal-widget-0.0.1...applications-portal-widget-0.0.2) (2024-09-29)
 
 ### Dependency Updates